MIPS: Backport generic div64 and delay. Since h __asm__ qualifier was deprecated in GCC 4.4 and later it became impossible to compile with later GCC versions. It was resolved with replacing __asm__ portions with generic c code.

+++ arch/mips/include/asm/delay.h
@@ -11,9 +11,11 @@
 #ifndef _ASM_DELAY_H
 #define _ASM_DELAY_H
 
-extern void __delay(unsigned int loops);
-extern void __ndelay(unsigned int ns);
-extern void __udelay(unsigned int us);
+#include <linux/param.h>
+
+extern void __delay(unsigned long loops);
+extern void __ndelay(unsigned long ns);
+extern void __udelay(unsigned long us);
 
 #define ndelay(ns) __ndelay(ns)
 #define udelay(us) __udelay(us)
